Fix tests by making use of template attribute inference when possible
Many times when the functions are already templates, manually annotating
with "nothrow" is not needed as it it inferred when possible. This is
proven with some unittests that are nothrow, but for the inference to
work, the tests have to use classes with their opEquals annotated
nothrow too.

Using a template without arguments `put` for attribute inference might
be a breaking change. It will interfere with client code that takes the
address of the function, but then it works in attributed code.

I don't know how the first test in MDList.d can be made `@safe`. It
stores pointers to a stack variable, I'm not sure if the compiler
supports this yet for safe code. I disabled the `@safe` attribute for
the test, this might be worth looking into later.